+++ Godeps/_workspace/src/github.com/drone/drone-plugin-go/plugin/const.go
@@ -1,0 +1,17 @@
+package plugin
+
+const (
+	StatePending = "pending"
+	StateRunning = "running"
+	StateSuccess = "success"
+	StateFailure = "failure"
+	StateKilled  = "killed"
+	StateError   = "error"
+)
+
+const (
+	EventPush   = "push"
+	EventPull   = "pull_request"
+	EventTag    = "tag"
+	EventDeploy = "deploy"
+)