+++ Godeps/_workspace/src/sourcegraph.com/sourcegraph/srclib/cli/repo_cmds.go
@@ -28,8 +28,6 @@ func (c *repoCmd) Execute(args []string) error {
 		return err
 	}
 	colorable.Println("# Current repository:")
-	colorable.Println("URI:", repo.URI())
-	colorable.Println("Clone URL:", repo.CloneURL)
 	colorable.Println("VCS:", repo.VCSType)
 	colorable.Println("Root dir:", repo.RootDir)
 	colorable.Println("Commit ID:", repo.CommitID)