+++ Godeps/_workspace/src/sourcegraph.com/sourcegraph/srclib/cli/store_cmds.go
@@ -68,7 +68,6 @@ func InitStoreCmds(c *flags.Command) {
 	if err != nil {
 		log.Fatal(err)
 	}
-	SetDefaultRepoOpt(importC)
 	SetDefaultCommitIDOpt(importC)
 
 	_, err = c.AddCommand("indexes",
@@ -235,7 +234,7 @@ func Import(buildDataFS vfs.FileSystem, stor interface{}, opt ImportOpt) error {
 	if err != nil {
 		return fmt.Errorf("error calling config.ReadCached: %s", err)
 	}
-	mf, err := plan.CreateMakefile(".", nil, "", treeConfig, plan.Options{})
+	mf, err := plan.CreateMakefile(".", nil, "", treeConfig)
 	if err != nil {
 		return fmt.Errorf("error calling plan.Makefile: %s", err)
 	}